+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **[ FEATURE ]** Add `json` subcommand that allows users to build
   programmatic extensions
 - **[ FEATURE ]** Support Windows line endings (`\r\n`)
+- **[ FEATURE ]** Add `bookmark unset` command for clearing current selection
+- **[ FEATURE ]** Check stdin for input (to allow shell piping)
 
 ## v1.5
 - **[ FIX ]** Fix the ongoing time counter in `klog now --follow`